+++ frontend/app/studio/[brainId]/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/PublicPrompts.module.scss
@@ -1,0 +1,4 @@
+.selector_wrapper {
+  width: 30%;
+  min-width: 250px;
+}
