--- frontend/app/studio/[brainId]/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/components/PublicPromptsList/index.ts
@@ -1,1 +1,0 @@
-export * from "./PublicPromptsList";
