--- frontend/app/studio/[brainId]/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/components/index.ts
@@ -1,1 +1,0 @@
-export * from "./PublicPromptsList";
