--- frontend/app/studio/[brainId]/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/index.ts
@@ -1,1 +1,0 @@
-export * from "./PublicPrompts";
