+++ frontend/lib/components/ui/SingleSelector/SingleSelector.module.scss
@@ -19,7 +19,7 @@
     cursor: pointer;
 
     &.unfolded {
-      border-radius: Radius.$normal Radius.$normal 0 0;
+      border-radius: Radius.$big Radius.$big 0 0;
     }
 
     .left {
